Re-added the timescale migrations for deployment to production
The message below is mostly the same as in the previous commit, but for future readers I added this explanation here as well.

In order to do the timescale migration, the receiving of the data should first be moved to the new fault tolerant endpoint. There have been adjustments to this system after the timescale migrations on acc though. For that reason this commit removes the migrations to timescale (and those thereafter). That way we can deploy this to production. Add the new fault tolerant endpoint. And then re-add these timescale migrations again once the data stream has been moved to the fault tolerant endpoint.