Add possible null values to db
Furthermore, this commit also includes a migration of the timestamp fields to timescale. This migration to timescale was previously done
in a manually written migration called 0002_to_timescale. Django hasn't officially did that change though so when running manage.py makemigrations
it also added these. These migrations don't do anything though, so it's more to keep the django records tidy.